feat(data table): refactor DataTable structure into compound components
